XRDDEV-1539 Update notifications
