+++ src/proxy-ui-api/frontend/src/views/Clients/Subsystem.vue
@@ -25,13 +25,6 @@
  -->
 <template>
   <div class="xrd-sub-view-wrapper">
-    <sub-tabs :tab="tab">
-      <v-tab v-for="tab in tabs" v-bind:key="tab.key" :to="tab.to">{{
-        $t(tab.name)
-      }}</v-tab>
-    </sub-tabs>
-    <alerts-container />
-
     <div class="content xrd-view-common">
       <v-flex mb-4 class="title-action">
         <div v-if="client" class="xrd-view-title mb-3">
@@ -55,19 +48,14 @@
 <script lang="ts">
 import Vue from 'vue';
 import { mapGetters } from 'vuex';
-import { Permissions, RouteName } from '@/global';
-import { Tab } from '@/ui-types';
-import SubTabs from '@/components/layout/SubTabs.vue';
+import { Permissions } from '@/global';
 import DeleteClientButton from '@/components/client/DeleteClientButton.vue';
 import UnregisterClientButton from '@/components/client/UnregisterClientButton.vue';
-import AlertsContainer from '@/components/ui/AlertsContainer.vue';
 
 export default Vue.extend({
   components: {
-    AlertsContainer,
     UnregisterClientButton,
     DeleteClientButton,
-    SubTabs,
   },
   props: {
     id: {
@@ -77,7 +65,6 @@ export default Vue.extend({
   },
   data() {
     return {
-      tab: undefined as undefined | Tab,
       confirmUnregisterClient: false as boolean,
       unregisterLoading: false as boolean,
     };
@@ -104,57 +91,6 @@ export default Vue.extend({
       }
 
       return this.$store.getters.hasPermission(Permissions.DELETE_CLIENT);
-    },
-
-    tabs(): Tab[] {
-      const allTabs: Tab[] = [
-        {
-          key: 'details',
-          name: 'tab.client.details',
-          to: {
-            name: RouteName.SubsystemDetails,
-            params: { id: this.id },
-          },
-        },
-        {
-          key: 'serviceClients',
-          name: 'tab.client.serviceClients',
-          to: {
-            name: RouteName.SubsystemServiceClients,
-            params: { id: this.id },
-          },
-          permissions: [Permissions.VIEW_CLIENT_ACL_SUBJECTS],
-        },
-        {
-          key: 'services',
-          name: 'tab.client.services',
-          to: {
-            name: RouteName.SubsystemServices,
-            params: { id: this.id },
-          },
-          permissions: [Permissions.VIEW_CLIENT_SERVICES],
-        },
-        {
-          key: 'internalServers',
-          name: 'tab.client.internalServers',
-          to: {
-            name: RouteName.SubsystemServers,
-            params: { id: this.id },
-          },
-          permissions: [Permissions.VIEW_CLIENT_INTERNAL_CERTS],
-        },
-        {
-          key: 'localGroups',
-          name: 'tab.client.localGroups',
-          to: {
-            name: RouteName.SubsystemLocalGroups,
-            params: { id: this.id },
-          },
-          permissions: [Permissions.VIEW_CLIENT_LOCAL_GROUPS],
-        },
-      ];
-
-      return this.$store.getters.getAllowedTabs(allTabs);
     },
   },
   created() {
